Extract repo get
