+++ Cargo.toml
@@ -17,3 +17,4 @@ regex = "1.3.1"
 reqwest = "0.9.10"
 json = "0.11.13"
 failure = "0.1.5"
+git2 = "0.11.0"
